Refactoring the notification hack. Changing the default of the notifications hack to not be enabled as most moder linux clients have started to support it. Introducing the Teams-for-linux version in the help menu option. Moving the handling of the login dialog to the login folder
