+++ app/index.js
@@ -1,13 +1,55 @@
-const { shell, app, ipcMain, BrowserWindow } = require('electron');
+const { shell, app, BrowserWindow } = require('electron');
 const windowStateKeeper = require('electron-window-state');
-const NativeNotification = require('electron-native-notification');
 const path = require('path');
 const iconPath = path.join(__dirname, 'assets', 'icons', 'icon-96x96.png');
 const config = require('./config')(app.getPath('userData'));
 const login = require('./login');
 const Menus = require('./menus');
-
+const notifications = require('./notifications');
+	
 global.edgeUserAgent = config.edgeUserAgent;
+
+app.commandLine.appendSwitch('auth-server-whitelist', config.authServerWhitelist);
+app.commandLine.appendSwitch('enable-ntlm-v2', config.ntlmV2enabled);
+
+app.on('ready', () => {
+	let window = createWindow();
+	new Menus(window, config, iconPath);
+
+	window.on('page-title-updated', (event, title) => {
+		window.webContents.send('page-title', title);
+	});
+
+	if (config.enableDesktopNotificationsHack) {
+		notifications.addDesktopNotificationHack(iconPath);
+	}
+
+	window.webContents.on('new-window', (event, url, frame, disposition) => {
+		if (disposition !== 'background-tab') {      
+			event.preventDefault();
+			shell.openExternal(url);
+		}
+	});
+
+	login.handleLoginDialogTries(window.webContents);
+	
+	window.webContents.setUserAgent(config.chromeUserAgent);
+
+	window.once('ready-to-show', () => window.show());
+
+	window.webContents.on('did-finish-load', () => {
+		window.webContents.insertCSS('#download-mobile-app-button, #download-app-button, #get-app-button { display:none; }');
+		window.webContents.insertCSS('.zoetrope { animation-iteration-count: 1 !important; }');
+	});
+
+	window.on('closed', () => { window = null; });
+
+	window.loadURL(config.url);
+
+	if (config.webDebug) {
+		window.openDevTools();
+	}
+});
 
 function createWindow() {
 	// Load the previous state with fallback to defaults
@@ -44,74 +86,3 @@ function createWindow() {
 
 	return window;
 }
-
-app.commandLine.appendSwitch('auth-server-whitelist', config.authServerWhitelist);
-app.commandLine.appendSwitch('enable-ntlm-v2', config.ntlmV2enabled);
-
-app.on('ready', () => {
-	let isFirstLoginTry = true;
-	let window = createWindow();
-	new Menus(window, config, iconPath);
-
-	window.on('page-title-updated', (event, title) => {
-		window.webContents.send('page-title', title);
-	});
-
-	if (!config.disableDesktopNotificationsHack) {
-		ipcMain.on('notifications', async (e, msg) => {
-			if (msg.count > 0) {
-				const body = ((msg.text) ? `(${msg.count}): ${msg.text}` : `You got ${msg.count} notification(s)`);
-				const notification = new NativeNotification(
-					'Microsoft Teams',
-					{
-						body,
-						icon: iconPath,
-					},
-				);
-				notification.onclick = () => {
-					window.show();
-					window.focus();
-				};
-				if (notification.show !== undefined) {
-					notification.show();
-				}
-			}
-		});
-	}
-
-	window.webContents.on('new-window', (event, url, frame, disposition) => {
-		if (disposition !== 'background-tab') {      
-			event.preventDefault();
-			shell.openExternal(url);
-		}
-	});
-
-	window.webContents.on('login', (event, request, authInfo, callback) => {
-		event.preventDefault();
-		if (isFirstLoginTry) {
-			isFirstLoginTry = false;
-			login.loginService(window, callback);
-		} else {
-			isFirstLoginTry = true;
-			app.relaunch();
-			app.exit(0);
-		}
-	});
-
-	window.webContents.setUserAgent(config.chromeUserAgent);
-
-	window.once('ready-to-show', () => window.show());
-
-	window.webContents.on('did-finish-load', () => {
-		window.webContents.insertCSS('#download-mobile-app-button, #download-app-button, #get-app-button { display:none; }');
-		window.webContents.insertCSS('.zoetrope { animation-iteration-count: 1 !important; }');
-	});
-
-	window.on('closed', () => { window = null; });
-
-	window.loadURL(config.url);
-
-	if (config.webDebug) {
-		window.openDevTools();
-	}
-});
