+++ app/login/index.js
@@ -1,4 +1,20 @@
-const {ipcMain, BrowserWindow} = require('electron');
+const {app, ipcMain, BrowserWindow} = require('electron');
+let isFirstLoginTry = true;
+
+exports.handleLoginDialogTries = function handleLoginDialogTries(webContents) {
+	webContents.on('login', (event, request, authInfo, callback) => {
+		event.preventDefault();
+		if (isFirstLoginTry) {
+			isFirstLoginTry = false;
+			this.loginService(window, callback);
+		} else {
+			// if fails to authenticate we need to relanch the app as we are closed 
+			isFirstLoginTry = true;
+			app.relaunch();
+			app.exit(0);
+		}
+	});
+};
 
 exports.loginService = function loginService(parentWindow, callback) {
 	let win = new BrowserWindow({
