+++ app/menus/help.js
@@ -1,19 +1,23 @@
-const open = require('opn');
+const {shell} = require('electron');
 
 exports = module.exports = (app) => ({
 	label: 'Help',
 	submenu: [
 		{
 			label: 'Online Documentation',
-			click: () => open('https://support.office.com/en-us/teams?omkt=en-001'),
+			click: () => shell.openExternal('https://support.office.com/en-us/teams?omkt=en-001'),
 		},
 		{
 			label: 'Github Project',
-			click: () => open('https://github.com/IsmaelMartinez/teams-for-linux'),
+			click: () => shell.openExternal('https://github.com/IsmaelMartinez/teams-for-linux'),
 		},
 		{type: 'separator'},
 		{
-			label: `Version ${app.getVersion()}`,
+			label: `Teams-for-linux version ${require('./../../package.json').version}`,
+			enabled: false,
+		},
+		{
+			label: `Electron version ${app.getVersion()}`,
 			enabled: false,
 		},
 		{role: 'toggledevtools'},
