+++ app/notifications/README.md
@@ -1,0 +1,7 @@
+This folder has the code that handles the desktop notifications hack implementation.
+
+There only the [index.js](index.js) file. The desktop notification is a temporary solution that will be hopefully face out as linux clients start implementing its support. 
+
+At the time of writing, Ubuntu 18 does support the notications so it has been deem necessary to disable this hack. 
+
+It is possible to enable it by using the configuration options. Check the [config README.md](../config/README.md) for more info on how to enableDesktopNotificationsHack
