+++ package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "electron-native-notification": "^1.2.1",
     "electron-window-state": "^4.1.1",
-    "opn": "^5.3.0",
     "yargs": "^12.0.1"
   },
   "devDependencies": {
